[Cleanup] Move `revealClosedDetailsAncestors` to a new file https://bugs.webkit.org/show_bug.cgi?id=295378 rdar://154912477
Reviewed by Aditya Keerthi.

I plan to add:
- a new algorithm for `hidden=until-found`
- shared code for auto-expanding details and `hidden=until-found`.

into that file.

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/dom/FindRevealAlgorithms.cpp: Added.
(WebCore::revealClosedDetailsAncestors):
* Source/WebCore/dom/FindRevealAlgorithms.h: Added.
* Source/WebCore/dom/TreeScope.h:
* Source/WebCore/html/HTMLDetailsElement.cpp:
(WebCore::revealClosedDetailsAncestors): Deleted.
* Source/WebCore/html/HTMLDetailsElement.h:
* Source/WebCore/page/LocalFrameView.cpp:
* Source/WebKit/WebProcess/WebPage/FindController.cpp:
* Source/WebKit/WebProcess/WebPage/WebFoundTextRangeController.cpp:
* Source/WebKit/WebProcess/WebPage/ios/FindControllerIOS.mm:

Canonical link: https://commits.webkit.org/296955@main
