+++ Source/WebCore/dom/TreeScope.h
@@ -28,10 +28,12 @@
 
 #include "HitTestSource.h"
 #include <memory>
+#include <wtf/CheckedRef.h>
 #include <wtf/Forward.h>
 #include <wtf/NoVirtualDestructorBase.h>
 #include <wtf/UniqueRef.h>
 #include <wtf/Vector.h>
+#include <wtf/WeakHashSet.h>
 #include <wtf/text/AtomString.h>
 
 namespace JSC {
