+++ Source/WebCore/page/LocalFrameView.cpp
@@ -49,6 +49,7 @@
 #include "EventHandler.h"
 #include "EventLoop.h"
 #include "EventNames.h"
+#include "FindRevealAlgorithms.h"
 #include "FixedContainerEdges.h"
 #include "FloatRect.h"
 #include "FocusController.h"
@@ -60,7 +61,6 @@
 #include "FrameTree.h"
 #include "GraphicsContext.h"
 #include "HTMLBodyElement.h"
-#include "HTMLDetailsElement.h"
 #include "HTMLEmbedElement.h"
 #include "HTMLFrameElement.h"
 #include "HTMLFrameSetElement.h"
